unifying video_ops and tweaking get_videos_in_dir
