+++ lightning_pose/utils/video_ops.py
@@ -6,6 +6,8 @@ from sklearn.decomposition import PCA
 from sklearn.cluster import KMeans
 import torch
 from tqdm import tqdm
+import subprocess
+from typing import List
 
 
 from lightning_pose.data.dali import video_pipe, LitDaliWrapper
@@ -188,3 +190,33 @@ def get_frames_from_idxs(cap, idxs):
             )
             break
     return frames
+
+
+def reencode_video(input_file: str, output_file: str) -> None:
+    """ a function that executes ffmpeg from a subprocess
+    reencodes video into H.264 coded format
+    input_file: str with abspath to existing video
+    outputfile: str with abspath to to new video"""
+    assert os.path.isfile(input_file), "input video does not exist." # input file exists
+    assert os.path.isdir(os.path.dirname(output_file)), "saving folder %s does not exist." % os.path.dirname(output_file) # folder for saving outputs exists
+    ffmpeg_cmd = f'ffmpeg -i {input_file} -c:v libx264 -c:a copy -y {output_file}'
+    subprocess.run(ffmpeg_cmd, shell=True)
+
+def check_codec_format(input_file: str):
+    # Run FFprobe command to get video codec version
+    # command = ["ffprobe", "-v", "error", "-select_streams", "v:0", "-show_entries", "stream=codec_version", "-of", "default=noprint_wrappers=1:nokey=1", input_file]
+    ffmpeg_cmd = f'ffmpeg -i {input_file}'
+    output_str = subprocess.run(ffmpeg_cmd, shell=True, capture_output=True, text=True)
+    output_str = output_str.stderr # stderr because the ffmpeg command has no output file. but the stderr still has codec info.
+
+    # search for h264
+    if output_str.find('h264') != -1:
+        # print('Video uses H.264 codec')
+        is_codec = True
+    else:
+        # print('Video does not use H.264 codec')
+        is_codec = False
+    return is_codec
+
+
+
